Made the app, PHP 7 compliant, moved the concrete implementation to services folder
