+++ app/Services/DefaultCalculation.php
@@ -1,5 +1,6 @@
 <?php
-namespace App\Repositories;
+declare(strict_types=1);
+namespace App\Services;
 
 use  App\Repositories\Contracts\CalculationInterface;
 
@@ -7,19 +8,19 @@ class DefaultCalculation implements CalculationInterface
 {
 	/**
 	 *This method does a calculation of 2 numbers
-	 *@return int
+	 *@return float
      *
     */
-	public function calculate($number1, $number2)
+	public function calculate(float $number1, float $number2):float
 	{
-		return '';
+		return 0.0;
 	}
 	/**
 	 *This method returns the type of operator 
 	 *@return string
      *
     */
-	public function getOperatorType()
+	public function getOperatorType():string
 	{
 		return '';
 	}
