+++ app/Services/DivisionCalculation.php
@@ -1,5 +1,6 @@
 <?php
-namespace App\Repositories;
+declare(strict_types=1);
+namespace App\Services;
 
 use  App\Repositories\Contracts\CalculationInterface;
 
@@ -7,10 +8,10 @@ class DivisionCalculation implements CalculationInterface
 {
 	/**
 	 *This method does an division calculation of 2 numbers
-	 *@return int
+	 *@return float
      *
     */
-	public function calculate($number1, $number2)
+	public function calculate(float $number1, float $number2):float
 	{
 		return($number1 / $number2);
 	}
@@ -19,7 +20,7 @@ class DivisionCalculation implements CalculationInterface
 	 *@return string
      *
     */
-	public function getOperatorType()
+	public function getOperatorType():string
 	{
 		return '/';
 	}
